Refactor live output services and UI
