+++ Sources/DependencyClients/SSHClient.swift
@@ -432,21 +432,6 @@ package enum SSHError: LocalizedError, Equatable {
   }
 }
 
-// Helper function to properly escape shell arguments
-private func escapeShellArgument(_ argument: String) -> String {
-  // If the argument contains any special characters, wrap it in single quotes
-  // and escape any single quotes within it
-  let specialChars = CharacterSet(charactersIn: " \t\n\r'\"\\$`;&|()<>*?[]{}!")
-
-  if argument.rangeOfCharacter(from: specialChars) != nil {
-    // Replace single quotes with '\'' (end quote, escaped quote, start quote)
-    let escaped = argument.replacingOccurrences(of: "'", with: "'\"'\"'")
-    return "'\(escaped)'"
-  }
-
-  return argument
-}
-
 // Helper function to extract detailed error information from NIOSSH errors
 private func detailedErrorDescription(_ error: Error) -> String {
   // Map ChannelError to friendly descriptions first
@@ -1331,93 +1316,6 @@ private final class CommandOutputHandler: ChannelInboundHandler, @unchecked Send
   }
 }
 
-// Streams stdout/stderr lines to callbacks; completes on exit or channel close.
-private final class LineStreamHandler: ChannelInboundHandler, @unchecked Sendable {
-  typealias InboundIn = SSHChannelData
-
-  private var stdoutBuffer = Data()
-  private var stderrBuffer = Data()
-  private var finished = false
-  private let onLine: @Sendable (String) -> Void
-  private let onErrorLine: @Sendable (String) -> Void
-  private let onFinish: @Sendable (Error?) -> Void
-
-  init(
-    onLine: @escaping @Sendable (String) -> Void,
-    onErrorLine: @escaping @Sendable (String) -> Void,
-    onFinish: @escaping @Sendable (Error?) -> Void
-  ) {
-    self.onLine = onLine
-    self.onErrorLine = onErrorLine
-    self.onFinish = onFinish
-  }
-
-  func channelRead(context: ChannelHandlerContext, data: NIOAny) {
-    let channelData = self.unwrapInboundIn(data)
-    switch channelData.type {
-    case .channel:
-      if case .byteBuffer(var buf) = channelData.data, let bytes = buf.readBytes(length: buf.readableBytes) {
-        stdoutBuffer.append(contentsOf: bytes)
-        flushLines(buffer: &stdoutBuffer, emit: onLine)
-      }
-    case .stdErr:
-      if case .byteBuffer(var buf) = channelData.data, let bytes = buf.readBytes(length: buf.readableBytes) {
-        stderrBuffer.append(contentsOf: bytes)
-        flushLines(buffer: &stderrBuffer, emit: onErrorLine)
-      }
-    default:
-      break
-    }
-  }
-
-  func userInboundEventTriggered(context: ChannelHandlerContext, event: Any) {
-    if let exitStatus = event as? SSHChannelRequestEvent.ExitStatus {
-      if !finished {
-        finished = true
-        if exitStatus.exitStatus == 0 {
-          // Flush any partial trailing lines
-          flushRemainder()
-          onFinish(nil)
-        } else {
-          let err = SSHError.commandFailed("Command finished with exit code: \(exitStatus.exitStatus)")
-          onFinish(err)
-        }
-      }
-    } else if event is SSHChannelRequestEvent.ExitSignal {
-      if !finished {
-        finished = true
-        flushRemainder()
-        onFinish(SSHError.commandFailed("Command terminated by signal"))
-      }
-    }
-  }
-
-  func channelInactive(context: ChannelHandlerContext) {
-    if !finished {
-      finished = true
-      flushRemainder()
-      onFinish(nil)
-    }
-  }
-
-  private func flushLines(buffer: inout Data, emit: (String) -> Void) {
-    while let range = buffer.firstRange(of: Data([0x0A])) { // '\n'
-      let lineData = buffer.subdata(in: buffer.startIndex..<range.lowerBound)
-      buffer.removeSubrange(buffer.startIndex..<range.upperBound) // remove including '\n'
-      if let line = String(data: lineData, encoding: .utf8) {
-        emit(line)
-      }
-    }
-  }
-
-  private func flushRemainder() {
-    if !stdoutBuffer.isEmpty, let line = String(data: stdoutBuffer, encoding: .utf8) { onLine(line) }
-    if !stderrBuffer.isEmpty, let line = String(data: stderrBuffer, encoding: .utf8) { onErrorLine(line) }
-    stdoutBuffer.removeAll(keepingCapacity: false)
-    stderrBuffer.removeAll(keepingCapacity: false)
-  }
-}
-
 private final class SSHUserAuthDelegate: NIOSSHClientUserAuthenticationDelegate, @unchecked Sendable {
   private let config: Models.SSHServerConfiguration
   private var authenticationAttempts = 0
@@ -1729,500 +1627,7 @@ package enum SSHConnectionError: Error, LocalizedError {
   }
 }
 
-struct TmuxService: Sendable {
-  private let config: Models.SSHServerConfiguration
-  private let sshClient: SSHClient
-
-  init(config: Models.SSHServerConfiguration) {
-    self.config = config
-    self.sshClient = SSHClient()
-  }
-
-  func hasSession(_ name: String) async throws -> Bool {
-    do {
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      return try await connectionManager.withConnection { connection in
-        let escapedName = escapeShellArgument(name)
-        let command = "tmux has-session -t \(escapedName) 2>/dev/null && echo 'exists' || echo 'not found'"
-        let result = try await connection.exec(command)
-        return result.trimmingCharacters(in: .whitespacesAndNewlines) == "exists"
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        throw SSHError.commandFailed(
-          "Tmux session check was cancelled. This may be due to network issues or server timeout.")
-      }
-
-      throw error
-    }
-  }
-
-  func newSession(name: String, path: String) async throws {
-    do {
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      try await connectionManager.withConnection { connection in
-        let escapedName = escapeShellArgument(name)
-        let escapedPath = escapeShellArgument(path)
-        let command = "tmux new-session -d -s \(escapedName) -c \(escapedPath)"
-        do {
-          _ = try await connection.exec(command)
-          await AppLogger.shared.log(
-            "Created tmux session: \(name) at \(path)",
-            level: .info,
-            category: .workspace
-          )
-        } catch {
-          // If the session already exists, tmux prints "duplicate session: <name>" to stderr with non-zero exit.
-          if case let SSHError.commandFailed(message) = error,
-            message.lowercased().contains("duplicate session") {
-            await AppLogger.shared.log(
-              "Using existing tmux session: \(name)",
-              level: .info,
-              category: .workspace
-            )
-            return
-          }
-          throw error
-        }
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        throw SSHError.commandFailed(
-          "Tmux session creation was cancelled. This may be due to network issues or server timeout."
-        )
-      }
-
-      throw error
-    }
-  }
-
-  func newOrReplaceServerWindow(name: String) async throws {
-    do {
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      try await connectionManager.withConnection { connection in
-        let hasExisting = try await hasSession(name)
-        if hasExisting {
-          let escapedName = escapeShellArgument(name)
-          let killCommand = "tmux kill-session -t \(escapedName)"
-          _ = try await connection.exec(killCommand)
-        }
-
-        let workspacePath = "$HOME"
-        try await newSession(name: name, path: workspacePath)
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        throw SSHError.commandFailed(
-          "Tmux server window setup was cancelled. This may be due to network issues or server timeout."
-        )
-      }
-
-      throw error
-    }
-  }
-
-  func listSessions() async throws -> [String] {
-    do {
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      return try await connectionManager.withConnection { connection in
-        let command = "tmux list-sessions -F '#{session_name}' 2>/dev/null || true"
-        let result = try await connection.exec(command)
-        return result.split(separator: "\n").map(String.init)
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        throw SSHError.commandFailed(
-          "Tmux session listing was cancelled. This may be due to network issues or server timeout."
-        )
-      }
-
-      throw error
-    }
-  }
-
-  func killSession(_ name: String) async throws {
-    do {
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      try await connectionManager.withConnection { connection in
-        let escapedName = escapeShellArgument(name)
-        let command = "tmux kill-session -t \(escapedName) 2>/dev/null || true"
-        _ = try await connection.exec(command)
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        throw SSHError.commandFailed(
-          "Tmux session kill was cancelled. This may be due to network issues or server timeout.")
-      }
-
-      throw error
-    }
-  }
-}
-
-// MARK: - tmux window/pane helpers
-extension TmuxService {
-  /// Ensure a window with the given name exists in the session, creating it if needed.
-  func ensureWindow(session: String, window: String, path: String) async throws {
-    let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-    try await connectionManager.withConnection { connection in
-      let escapedSession = escapeShellArgument(session)
-      let escapedWindow = escapeShellArgument(window)
-      let escapedPath = escapeShellArgument(path)
-
-      // If window exists, nothing to do
-      let listCommand = "tmux list-windows -t \(escapedSession) -F '#{window_name}' 2>/dev/null || true"
-      let windowsList = try await connection.exec(listCommand)
-      let names = windowsList.split(separator: "\n").map(String.init)
-      if names.contains(where: { $0 == window }) {
-        return
-      }
-
-      // Create a new window in the session
-      let newWindowCmd = "tmux new-window -t \(escapedSession): -n \(escapedWindow) -c \(escapedPath)"
-      do {
-        _ = try await connection.exec(newWindowCmd)
-        await AppLogger.shared.log(
-          "Created tmux window '\(window)' in session '\(session)'",
-          level: .info,
-          category: .workspace
-        )
-      } catch {
-        // If duplicate window due to race, ignore
-        if case let SSHError.commandFailed(message) = error,
-           message.lowercased().contains("duplicate window") || message.lowercased().contains("duplicate") {
-          return
-        }
-        throw error
-      }
-    }
-  }
-
-  /// Respawn the first pane in the named window with the provided command and working directory.
-  func respawnPane(session: String, window: String, path: String, command: String) async throws {
-    let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-    try await connectionManager.withConnection { connection in
-      let escapedSession = escapeShellArgument(session)
-      let escapedWindow = escapeShellArgument(window)
-      let escapedPath = escapeShellArgument(path)
-      // Safely single-quote the command for tmux invocation
-      let quoted = command.replacingOccurrences(of: "'", with: "'\"'\"'")
-      let respawnCmd = "tmux respawn-pane -k -c \(escapedPath) -t \(escapedSession):\(escapedWindow).0 '\(quoted)'"
-      _ = try await connection.exec(respawnCmd)
-      await AppLogger.shared.log(
-        "Respawned tmux pane in '\(session):\(window)'",
-        level: .info,
-        category: .workspace
-      )
-    }
-  }
-}
-
-package struct WorkspaceService: Sendable {
-  private let config: Models.SSHServerConfiguration
-  private let tmuxService: TmuxService
-  private let sshClient: SSHClient
-
-  package init(config: Models.SSHServerConfiguration) {
-    self.config = config
-    self.sshClient = SSHClient()
-    self.tmuxService = TmuxService(config: config)
-  }
-
-  // Establish an SSH connection and ensure the tmux session exists for the workspace, nothing else.
-  // This intentionally avoids spawning the opencode server, parsing ports, or writing daemon files.
-  package func connectAndEnsureTmux(workspace: Models.Workspace) async throws {
-    await AppLogger.shared.log(
-      "Connecting and ensuring tmux session for: \(workspace.name)",
-      level: .info,
-      category: .workspace
-    )
-
-    let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-    try await connectionManager.withConnection { _ in
-      // Create-or-use existing session without a separate has-session check
-      try await tmuxService.newSession(name: workspace.tmuxSession, path: workspace.remotePath)
-    }
-  }
-
-  package struct SpawnResult: Equatable {
-    package let port: Int
-    package let online: Bool
-    package let error: SSHError?
-
-    // Explicit package initializer so other modules in the package can construct results
-    package init(port: Int, online: Bool, error: SSHError?) {
-      self.port = port
-      self.online = online
-      self.error = error
-    }
-  }
-
-  // swiftlint:disable:next function_body_length
-  package func attachOrSpawn(workspace: Models.Workspace) async throws -> SpawnResult {
-    await AppLogger.shared.log(
-      "Attaching or spawning workspace: \(workspace.name)", level: .info, category: .workspace)
-
-    do {
-      // Use connection manager for all SSH operations to reuse the same connection
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      return try await connectionManager.withConnection { connection in
-        // Step 1: Ensure tmux session exists (no pre-check; tolerate duplicates)
-        try await tmuxService.newSession(name: workspace.tmuxSession, path: workspace.remotePath)
-
-        let stateDirectory = workspaceStateDirectory(for: workspace)
-        let daemonPath = workspaceDaemonPath(for: workspace)
-        let logPath = workspaceLogPath(for: workspace)
-
-        // Ensure directory exists for daemon/log files
-        _ = try await connection.exec("mkdir -p \(stateDirectory)")
-
-        // Step 2: Check for existing daemon.json using the same connection
-        let checkCommand = "if [ -f \(daemonPath) ]; then cat \(daemonPath); else echo '{}'; fi"
-        let daemonContent = try await connection.exec(checkCommand)
-
-        // Parse daemon.json to check for existing port
-        let decoder = JSONDecoder()
-        if let data = daemonContent.data(using: String.Encoding.utf8),
-          let daemonInfo = try? decoder.decode([String: Int].self, from: data),
-          let existingPort = daemonInfo["port"] {
-
-          // Health probe existing port
-          await AppLogger.shared.log(
-            "Found existing daemon on port \(existingPort), checking health",
-            level: .info,
-            category: .workspace
-          )
-          if await healthCheck(port: existingPort, workspace: workspace) {
-            await AppLogger.shared.log(
-              "Existing daemon is healthy on port \(existingPort)",
-              level: .info,
-              category: .workspace
-            )
-            return SpawnResult(port: existingPort, online: true, error: nil)
-          } else {
-            await AppLogger.shared.log(
-              "Existing daemon is unhealthy on port \(existingPort), will spawn new instance",
-              level: .warning,
-              category: .workspace
-            )
-          }
-        }
-
-        // Step 3: Spawn opencode server with automatic port selection
-        await AppLogger.shared.log(
-          "Spawning opencode server for workspace: \(workspace.name)",
-          level: .info,
-          category: .workspace
-        )
-        // Build the command using NPX to run opencode-ai and print logs to shared state directory.
-        let serveCommand =
-          "npx -y opencode-ai serve --hostname 127.0.0.1 --port 0 --print-logs"
-        let spawnCommand =
-          "mkdir -p \(stateDirectory) && \(serveCommand) | tee -a \(logPath)"
-
-        // Ensure a dedicated tmux window named 'opencode' exists and respawn the pane with our command
-        try await tmuxService.ensureWindow(
-          session: workspace.tmuxSession,
-          window: "opencode",
-          path: workspace.remotePath
-        )
-
-        try await tmuxService.respawnPane(
-          session: workspace.tmuxSession,
-          window: "opencode",
-          path: workspace.remotePath,
-          command: spawnCommand
-        )
-
-        // Step 4: Wait for server to start and parse the assigned port from logs
-        let maxRetries = 30  // 30 seconds timeout
-        for _ in 0..<maxRetries {
-          if let assignedPort = try await parsePortFromLogs(workspace: workspace, connection: connection) {
-            // Create daemon.json with the actual assigned port
-            let daemonData = try JSONEncoder().encode(["port": assignedPort])
-            if let daemonJson = String(data: daemonData, encoding: .utf8) {
-              let writeCommand = "mkdir -p \(stateDirectory) && echo '\(daemonJson)' > \(daemonPath)"
-              _ = try await connection.exec(writeCommand)
-            }
-
-            if await healthCheck(port: assignedPort, workspace: workspace) {
-              await AppLogger.shared.log(
-                "OpenCode server started successfully on port \(assignedPort)",
-                level: .info,
-                category: .workspace
-              )
-              return SpawnResult(port: assignedPort, online: true, error: nil)
-            }
-          }
-          try await Task.sleep(for: .seconds(1))
-        }
-
-        await AppLogger.shared.log(
-          "Failed to start opencode server within timeout for workspace: \(workspace.name)",
-          level: .error,
-          category: .workspace
-        )
-        let timeoutError = SSHError.spawnTimeout("Failed to start opencode server within timeout")
-        return SpawnResult(port: 0, online: false, error: timeoutError)
-      }
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        let cancellationError = SSHError.spawnTimeout(
-          "Workspace spawn was cancelled. This may be due to network issues or server timeout.")
-        await AppLogger.shared.log(
-          "Workspace spawn cancelled for: \(workspace.name)",
-          level: .error,
-          category: .workspace
-        )
-        return SpawnResult(port: 0, online: false, error: cancellationError)
-      }
-
-      throw error
-    }
-  }
-
-  private func parsePortFromLogs(workspace: Models.Workspace, connection: SSHConnection) async throws -> Int? {
-    do {
-      // Read the live log (stored outside the workspace) to find the assigned port
-      let logPath = workspaceLogPath(for: workspace)
-      let command = "tail -n 50 \(logPath) 2>/dev/null || echo ''"
-      let logContent = try await connection.exec(command)
-
-      // Look for opencode server startup pattern (support opencode/opencode-ai variants):
-      // e.g., "opencode server listening on http://127.0.0.1:51535"
-      let pattern = #"(?i)(opencode|opencode\s*ai|opencode-ai).*server listening on http://[^:]+:(\d+)"#
-
-      if let regex = try? NSRegularExpression(pattern: pattern, options: .caseInsensitive),
-        let match = regex.firstMatch(
-          in: logContent, range: NSRange(logContent.startIndex..., in: logContent)),
-        let portRange = Range(match.range(at: 2), in: logContent) {
-        let portString = String(logContent[portRange])
-        if let port = Int(portString) {
-          return port
-        }
-      }
-
-      return nil
-    } catch {
-      // Handle CancellationError specifically - don't throw, just return nil
-      if error is CancellationError {
-        await AppLogger.shared.log(
-          "Port parsing was cancelled for workspace: \(workspace.name)",
-          level: .warning,
-          category: .workspace
-        )
-        return nil
-      }
-
-      throw error
-    }
-  }
-
-  private func healthCheck(port: Int, workspace: Models.Workspace) async -> Bool {
-    // Mock health check - in real implementation would probe the server
-    return port > 0
-  }
-
-  package func getLiveOutputStream(workspace: Models.Workspace) -> AsyncStream<String> {
-    return AsyncStream { continuation in
-      // Run the streaming setup in a Task
-      Task {
-        do {
-          let manager = await SSHConnectionPool.shared.manager(for: self.config)
-          try await manager.withConnection { connection in
-            let logPath = workspaceLogPath(for: workspace)
-            let command = "tail -n 200 -F \(logPath) 2>/dev/null"
-
-            let sessionPromise = connection.channel.eventLoop.makePromise(of: Channel.self)
-            let creationFuture: EventLoopFuture<Void> = connection.channel.eventLoop.submit {
-              let sshHandler = try connection.channel.pipeline.syncOperations.handler(type: NIOSSHHandler.self)
-              sshHandler.createChannel(sessionPromise, channelType: .session) { child, _ in
-                let handler = LineStreamHandler(
-                  onLine: { continuation.yield($0) },
-                  onErrorLine: { continuation.yield($0) },
-                  onFinish: { _ in continuation.finish() }
-                )
-                return child.pipeline.addHandler(handler)
-              }
-            }
-            try await creationFuture.get()
-            let streamChannel = try await sessionPromise.futureResult.get()
-
-            let exec = SSHChannelRequestEvent.ExecRequest(command: command, wantReply: true)
-            let noPromise: EventLoopPromise<Void>? = nil
-            streamChannel.triggerUserOutboundEvent(exec, promise: noPromise)
-
-            // Set termination handler after we have the channel
-            continuation.onTermination = { _ in
-              Task { try? await streamChannel.close().get() }
-            }
-          }
-        } catch {
-          continuation.finish()
-        }
-      }
-    }
-  }
-
-  package func cleanAndRetry(workspace: Models.Workspace) async throws -> SpawnResult {
-    do {
-      // Use connection manager for cleanup operations too
-      let connectionManager = await SSHConnectionPool.shared.manager(for: config)
-      try await connectionManager.withConnection { connection in
-        let stateDirectory = workspaceStateDirectory(for: workspace)
-        let daemonPath = workspaceDaemonPath(for: workspace)
-        let logPath = workspaceLogPath(for: workspace)
-        // Remove stale daemon.json, log, and lock files
-        let cleanupCommand = "rm -f \(daemonPath) \(logPath) \(stateDirectory)/lock"
-        _ = try await connection.exec(cleanupCommand)
-
-        // Kill existing tmux session
-        try await tmuxService.killSession(workspace.tmuxSession)
-      }
-
-      // Retry spawn
-      return try await attachOrSpawn(workspace: workspace)
-    } catch {
-      // Handle CancellationError specifically
-      if error is CancellationError {
-        let cancellationError = SSHError.spawnTimeout(
-          "Workspace cleanup and retry was cancelled. This may be due to network issues or server timeout."
-        )
-        await AppLogger.shared.log(
-          "Workspace cleanup cancelled for: \(workspace.name)",
-          level: .error,
-          category: .workspace
-        )
-        return SpawnResult(port: 0, online: false, error: cancellationError)
-      }
-
-      throw error
-    }
-  }
-}
-
-// MARK: - Workspace metadata paths
-
-extension WorkspaceService {
-  private func workspaceStateDirectory(for workspace: Models.Workspace) -> String {
-    "$HOME/.opencoder/workspaces/\(workspace.id.uuidString)"
-  }
-
-  private func workspaceDaemonPath(for workspace: Models.Workspace) -> String {
-    "\(workspaceStateDirectory(for: workspace))/daemon.json"
-  }
-
-  private func workspaceLogPath(for workspace: Models.Workspace) -> String {
-    "\(workspaceStateDirectory(for: workspace))/live.log"
-  }
-}
-
+// MARK: - Connection Management
 // MARK: - Connection Management
 
 package actor SSHConnectionManager {
