+++ Sources/Features/WorkspaceLogs.swift
@@ -2,10 +2,21 @@ import Foundation
 import Models
 import DependencyClients
 
+package enum WorkspaceLogsError: Error, LocalizedError, Equatable {
+  case missingConfiguration
+
+  package var errorDescription: String? {
+    switch self {
+    case .missingConfiguration:
+      return "No SSH server configuration linked to this workspace."
+    }
+  }
+}
+
 package enum WorkspaceLogs {
-  /// Returns a live AsyncStream of lines from the remote workspace's live log.
+  /// Returns a live AsyncStream of lines from the remote workspace's live log or tmux window.
   /// Falls back to an empty stream if no SSH configuration is linked to the workspace.
-  package static func stream(for workspace: Workspace) -> AsyncStream<String> {
+  package static func stream(for workspace: Workspace, window: String? = nil) -> AsyncStream<String> {
     guard let config = WorkspacesStorage.loadSSHConfigForWorkspace(workspace) else {
       return AsyncStream { continuation in
         continuation.yield("[Live Output] No SSH server configuration linked to this workspace.")
@@ -14,6 +25,16 @@ package enum WorkspaceLogs {
     }
 
     let service = WorkspaceService(config: config)
-    return service.getLiveOutputStream(workspace: workspace)
+    return service.getLiveOutputStream(workspace: workspace, window: window)
+  }
+
+  /// Returns the available tmux window names for the workspace's session.
+  package static func tmuxWindows(for workspace: Workspace) async throws -> [String] {
+    guard let config = WorkspacesStorage.loadSSHConfigForWorkspace(workspace) else {
+      throw WorkspaceLogsError.missingConfiguration
+    }
+
+    let service = WorkspaceService(config: config)
+    return try await service.listTmuxWindows(workspace: workspace)
   }
 }
